Added setfollowerwarppos macro and functionality
Before doing a warp, use setfollowerwarppos to determine how where the follower will appear relative to the player after the warp:
DIR_NORTH/SOUTH/EAST/WEST to place the follower above/below/right/left of the player. Without using other scripts to turn the follower, the follower will always face south after the warp.

DIR_NONE will cause the follower to be hidden and pop out of a pokeball as the player takes a step after the warp.